Split up trigger plugins
